health: #100 spread expansion code across files and remove index.ts in favor of public.ts - just something i'm trying out
