+++ src/expansion/copy-expansion.fn.ts
@@ -1,0 +1,6 @@
+import { Expansion } from "./expansion";
+import { mergeExpansions } from "./merge-expansions.fn";
+
+export function copyExpansion(expansion: Expansion): Expansion {
+    return mergeExpansions(expansion, {});
+}
