+++ src/expansion/public.ts
@@ -1,0 +1,4 @@
+export * from "./copy-expansion.fn";
+export * from "./expansion";
+export * from "./merge-expansions.fn";
+export * from "./reduce-expansion.fn";
