+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./criteria";
-export * from "./expansion";
+export * from "./expansion/public";
 export * from "./utils";
 export * from "./query";
 export * from "./model/property";
